feat (Operator) : OperatorGroup 클래스 이름변경, 문자열이 Operator 인지 확인하는 메소드 추가
